removing getuser method and moving methods to helpers
